alpha 6 : new query function, integrated previous basic functions : fullscreen, info, delete, inputToggle
