+++ index.html
@@ -54,15 +54,16 @@
   <div id="playerContainer">
     <div id="playbackButtons">
       <button class="material-icons on" id="playButton">play_arrow</button>
-      <button class="material-icons" id="queueButton">add_to_queue</button>
-      <button class="material-icons" id="nextButton">skip_next</button>
-      <button class="material-icons" id="settingsButton">settings</button>
+      <button class="material-icons   " id="queueButton">add_to_queue</button>
+      <button class="material-icons   " id="nextButton">skip_next</button>
+      <button class="material-icons   " id="settingsButton">settings</button>
     </div>
     <input type="url" placeholder="Input Link Here">
   </div>
-
+  <script src="JS/query.js"></script>
   <script src="JS/color.js"></script>
-  <script src="themer.js"></script>
+  <script src="JS/basics.js"></script>
+  <script src="JS/themer.js"></script>
 </body>
 
 </html>
